Move dev to runnable scripts
